requirements.txt
streamlit
pandas
plotly
gdown
duckdb
langchain
langchain-community
langchain-openai
langchain-experimental
python-dotenv
deep-translator
